First release (v0.52)
Files reshuffled and renamed for release.

Added __init__.py for release.

Added missing documentation.

Added license in each file.
