--- gui.py
@@ -1,6 +1,0 @@
-# The bhaiUI
-
-import tkinter
-
-homescreen = tkinter.Tk()
-canvas = tkinter.canvas(homescreen, width
